# docs/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "重生之凡人修仙传 API",
    "description": "修仙文字游戏后端接口文档，基于《凡人修仙传》世界观。包含玩家注册登录、角色管理、境界突破、地图探索、闭关修炼、GM管理、系统通知、WebSocket实时推送、玩家数据实时同步等完整功能。",
    "version": "2.2.0",
    "contact": {
      "name": "Shell",
      "email": "1592363624@qq.com"
    }
  },
  "servers": [
    {
      "url": "http://localhost:5000",
      "description": "本地开发环境"
    },
    {
      "url": "https://xx.52shell.ltd",
      "description": "生产环境"
    }
  ],
...
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT Token 认证"
      }
    },
...
  },
  "paths": {
...
    "/api/map/info": {
      "get": {
        "summary": "获取当前地图信息",
        "description": "获取玩家当前所在地图信息，合并静态配置和玩家动态数据",
        "tags": ["Map"],
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer",
                      "example": 200
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "current_map": {
                          "type": "object",
                          "description": "当前地图信息（静态配置 + 玩家动态数据）"
                        },
                        "connected_maps": {
                          "type": "array",
                          "description": "可连接的相邻地图列表（仅静态配置）"
                        }
                      }
                    }
                  }
                },
                "examples": {
                  "normal": {
                    "value": {
                      "code": 200,
                      "data": {
                        "current_map": {
                          "id": 1,
                          "name": "越国",
                          "type": "country",
                          "environment": "plains",
                          "required_realm": "凡人",
                          "danger_level": 1,
                          "travel_time": 5,
                          "description": "位于天南地区的凡人国度，韩立出身之地",
                          "npcs": [
                            { "id": "elder_zhang", "name": "张铁", "role": "npc", "location": "七玄门" }
                          ],
                          "resources": [
                            { "id": "firewood", "name": "干柴", "type": "gathering", "difficulty": 1, "yield": [1, 3] }
                          ],
                          "monsters": [],
                          "connections": [2],
                          "player_data": {
                            "exploration_progress": 25,
                            "visited_nodes": ["village_center", "north_gate"],
                            "last_visit_time": "2024-01-15T10:30:00.000Z"
                          }
                        },
                        "connected_maps": [
                          {
                            "id": 2,
                            "name": "七玄门",
                            "type": "sect",
                            "environment": "mountain",
                            "required_realm": "凡人",
                            "danger_level": 2,
                            "description": "越国的一个小门派，有一定修仙传承",
                            "travel_time": 3
                          }
                        ]
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/map/start-move": {
      "post": {
        "summary": "开始移动（时间移动系统）",
        "description": "开始移动到目标地图，进入移动状态。移动过程中玩家无法执行其他操作，移动完成后自动到达目标地点。",
        "tags": ["Map"],
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["targetMapId"],
                "properties": {
                  "targetMapId": {
                    "type": "integer",
                    "description": "目标地图ID"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "开始移动成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer",
                      "example": 200
                    },
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "message": {
                      "type": "string",
                      "example": "开始前往七玄门"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "from_map_id": {
                          "type": "integer",
                          "description": "出发地图ID"
                        },
                        "from_map_name": {
                          "type": "string",
                          "description": "出发地图名称"
                        },
                        "to_map_id": {
                          "type": "integer",
                          "description": "目标地图ID"
                        },
                        "to_map_name": {
                          "type": "string",
                          "description": "目标地图名称"
                        },
                        "distance": {
                          "type": "number",
                          "description": "移动距离（里）"
                        },
                        "mp_consumed": {
                          "type": "integer",
                          "description": "消耗的灵力"
                        },
                        "total_seconds": {
                          "type": "integer",
                          "description": "总耗时（秒）"
                        },
                        "start_time": {
                          "type": "string",
                          "format": "date-time",
                          "description": "移动开始时间"
                        },
                        "end_time": {
                          "type": "string",
                          "format": "date-time",
                          "description": "预计到达时间"
                        }
                      }
                    }
                  }
                },
                "examples": {
                  "normal": {
                    "value": {
                      "code": 200,
                      "success": true,
                      "message": "开始前往七玄门",
                      "data": {
                        "from_map_id": 1,
                        "from_map_name": "越国",
                        "to_map_id": 2,
                        "to_map_name": "七玄门",
                        "distance": 15.5,
                        "mp_consumed": 25,
                        "total_seconds": 93,
                        "start_time": "2024-05-20T10:00:00.000Z",
                        "end_time": "2024-05-20T10:01:33.000Z"
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "无法开始移动",
            "content": {
              "application/json": {
                "examples": {
                  "already_moving": {
                    "value": {
                      "code": 400,
                      "message": "正在移动中，无法重复移动"
                    }
                  },
                  "same_map": {
                    "value": {
                      "code": 400,
                      "message": "目标地图与当前地图相同"
                    }
                  },
                  "no_mp": {
                    "value": {
                      "code": 400,
                      "message": "灵力不足，需要 25 点灵力"
                    }
                  },
                  "realm_insufficient": {
                    "value": {
                      "code": 400,
                      "message": "境界不足，需达到 筑基期 才能进入此地图"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/map/cancel-move": {
      "post": {
        "summary": "取消移动",
        "description": "取消当前正在进行的移动，取消后返回出发地图，已消耗的灵力不退还",
        "tags": ["Map"],
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "取消成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer",
                      "example": 200
                    },
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "message": {
                      "type": "string",
                      "example": "已取消移动，返回越国"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "from_map_id": {
                          "type": "integer",
                          "description": "返回的地图ID"
                        },
                        "from_map_name": {
                          "type": "string",
                          "description": "返回的地图名称"
                        },
                        "mp_consumed": {
                          "type": "integer",
                          "description": "已消耗的灵力"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "无法取消",
            "content": {
              "application/json": {
                "example": {
                  "code": 400,
                  "message": "当前未在移动中"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/map/validate": {
      "get": {
        "summary": "验证地图配置",
        "description": "验证 map_data.json 配置的完整性，需要管理员权限",
        "tags": ["Map", "GM"],
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "验证结果",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "valid": { "type": "boolean", "description": "配置是否有效" },
                    "errors": { "type": "array", "items": { "type": "string" }, "description": "错误列表" }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}